rename: 네이밍 통일 및 불필요한 파일 제거
