+++ apps/frontend/src/pages/room/hooks/socket/index.ts
@@ -1,4 +1,4 @@
 export { useRoomSocketCache } from './useRoomSocketCache'
 export { useVoteSocket } from './useVoteSocket'
-export { useYjsSocket } from './yjs'
+export { useYjsSocket } from './useYjsSocket'
 export { useRoomSocket } from './useRoomSocket'
