+++ apps/frontend/src/pages/room/hooks/socket/yjs/index.ts
@@ -1,1 +1,5 @@
-export { useYjsSocket } from './useYjsSocket'
+export { useYjsDoc } from './useYjsDoc'
+export { useYjsHistory } from './useYjsHistory'
+export { useYjsSocketEvents } from './useYjsSocketEvents'
+export { useYjsTelemetry } from './useYjsTelemetry'
+export { useYjsCommands } from './useYjsCommands'
