+++ apps/frontend/src/pages/room/hooks/socket/yjs/useYjsCommands.ts
@@ -1,5 +1,5 @@
 import { useCallback } from 'react'
-import * as Y from 'yjs'
+import { Map as YMap, Doc as YDoc, UndoManager as YUndoManager } from 'yjs'
 import { YJS_TYPE, CANVAS_ITEM_TYPE, type CanvasItemType, type Line, type PlaceCard, type PostIt, type TextBox } from '@/shared/types'
 import { PLACE_CARD_HEIGHT, PLACE_CARD_WIDTH } from '@/pages/room/constants'
 import type { YjsRank } from '@/pages/room/types'
@@ -7,21 +7,21 @@ import { assignNextRank, makeKey, shouldSkipMoveToTop } from '@/pages/room/utils
 
 type RefValue<T> = { current: T }
 
-interface UseCanvasCommandsOptions {
-  docRef: RefValue<Y.Doc | null>
-  undoManagerRef: RefValue<Y.UndoManager | null>
+interface UseYjsCommandsProps {
+  docRef: RefValue<YDoc | null>
+  undoManagerRef: RefValue<YUndoManager | null>
   localOriginRef: RefValue<unknown>
   localMaxTimestampRef: RefValue<number>
   updateHistoryState: () => void
 }
 
-export const useCanvasCommands = ({ docRef, undoManagerRef, localOriginRef, localMaxTimestampRef, updateHistoryState }: UseCanvasCommandsOptions) => {
+export const useYjsCommands = ({ docRef, undoManagerRef, localOriginRef, localMaxTimestampRef, updateHistoryState }: UseYjsCommandsProps) => {
   const updateItem = useCallback(
     (canvasItemType: CanvasItemType, id: string, updates: Record<string, unknown>) => {
       const doc = docRef.current
       if (!doc) return
 
-      const yArray = doc.getArray<Y.Map<unknown>>(YJS_TYPE[canvasItemType])
+      const yArray = doc.getArray<YMap<unknown>>(YJS_TYPE[canvasItemType])
       const idToIndexMap = new Map<string, number>()
       yArray.forEach((yMap, index) => {
         const itemId = yMap.get('id') as string
@@ -48,9 +48,9 @@ export const useCanvasCommands = ({ docRef, undoManagerRef, localOriginRef, loca
       const doc = docRef.current
       if (!doc) return
 
-      const yPostits = doc.getArray<Y.Map<unknown>>(YJS_TYPE[CANVAS_ITEM_TYPE.POST_IT])
+      const yPostits = doc.getArray<YMap<unknown>>(YJS_TYPE[CANVAS_ITEM_TYPE.POST_IT])
       const yZRankByKey = doc.getMap<YjsRank>(YJS_TYPE.Z_RANK_BY_KEY)
-      const yMap = new Y.Map()
+      const yMap = new YMap()
       yMap.set('id', postit.id)
       yMap.set('x', postit.x)
       yMap.set('y', postit.y)
@@ -86,9 +86,9 @@ export const useCanvasCommands = ({ docRef, undoManagerRef, localOriginRef, loca
       const doc = docRef.current
       if (!doc) return
 
-      const yPlaceCards = doc.getArray<Y.Map<unknown>>(YJS_TYPE[CANVAS_ITEM_TYPE.PLACE_CARD])
+      const yPlaceCards = doc.getArray<YMap<unknown>>(YJS_TYPE[CANVAS_ITEM_TYPE.PLACE_CARD])
       const yZRankByKey = doc.getMap<YjsRank>(YJS_TYPE.Z_RANK_BY_KEY)
-      const yMap = new Y.Map()
+      const yMap = new YMap()
       yMap.set('id', card.id)
       yMap.set('placeId', card.placeId)
       yMap.set('name', card.name)
@@ -129,9 +129,9 @@ export const useCanvasCommands = ({ docRef, undoManagerRef, localOriginRef, loca
       const doc = docRef.current
       if (!doc) return
 
-      const yLines = doc.getArray<Y.Map<unknown>>(YJS_TYPE[CANVAS_ITEM_TYPE.LINE])
+      const yLines = doc.getArray<YMap<unknown>>(YJS_TYPE[CANVAS_ITEM_TYPE.LINE])
       const yZRankByKey = doc.getMap<YjsRank>(YJS_TYPE.Z_RANK_BY_KEY)
-      const yMap = new Y.Map()
+      const yMap = new YMap()
       yMap.set('id', line.id)
       yMap.set('points', line.points)
       yMap.set('stroke', line.stroke)
@@ -166,9 +166,9 @@ export const useCanvasCommands = ({ docRef, undoManagerRef, localOriginRef, loca
       const doc = docRef.current
       if (!doc) return
 
-      const yTextBoxes = doc.getArray<Y.Map<unknown>>(YJS_TYPE[CANVAS_ITEM_TYPE.TEXT_BOX])
+      const yTextBoxes = doc.getArray<YMap<unknown>>(YJS_TYPE[CANVAS_ITEM_TYPE.TEXT_BOX])
       const yZRankByKey = doc.getMap<YjsRank>(YJS_TYPE.Z_RANK_BY_KEY)
-      const yMap = new Y.Map()
+      const yMap = new YMap()
 
       Object.entries(textBox).forEach(([key, value]) => {
         yMap.set(key, value)
@@ -199,7 +199,7 @@ export const useCanvasCommands = ({ docRef, undoManagerRef, localOriginRef, loca
       const doc = docRef.current
       if (!doc) return
 
-      const yArray = doc.getArray<Y.Map<unknown>>(YJS_TYPE[canvasItemType])
+      const yArray = doc.getArray<YMap<unknown>>(YJS_TYPE[canvasItemType])
       const yZRankByKey = doc.getMap<YjsRank>(YJS_TYPE.Z_RANK_BY_KEY)
 
       const idToIndexMap = new Map<string, number>()
